8march code

+++ app/src/main/java/com/v2infotech/android/tiktok/database/DbHelper.java
@@ -11,16 +11,22 @@ import com.v2infotech.android.tiktok.Utils.Contants;
 import com.v2infotech.android.tiktok.model.LoginResponseData;
 import com.v2infotech.android.tiktok.model.UserProfileResponse;
 
-import static com.v2infotech.android.tiktok.Utils.Contants.IMAGE_PROFILE;
-import static com.v2infotech.android.tiktok.Utils.Contants.IMAGE_VIDEO;
-import static com.v2infotech.android.tiktok.Utils.Contants.TIKTOK_ID;
-import static com.v2infotech.android.tiktok.Utils.Contants.TIKTOK_USER_BIO;
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.v2infotech.android.tiktok.Utils.Contants.TIKTOK_CONTACT;
+import static com.v2infotech.android.tiktok.Utils.Contants.TIKTOK_EMAIL;
 import static com.v2infotech.android.tiktok.Utils.Contants.TIKTOK_USER_NAME;
+import static com.v2infotech.android.tiktok.Utils.Contants.USER_BIO;
+import static com.v2infotech.android.tiktok.Utils.Contants.USER_PROFILE_PIC;
 
 
 public class DbHelper extends SQLiteOpenHelper {
     // If you change the bl.womanguardian.database schema, you must increment the bl.womanguardian.database version.
-    public static final int DATABASE_VERSION =1;
+    Context ctx;
+    String DBNAME;
+    String DBPATH;
+    public static final int DATABASE_VERSION = 1;
     public static final String DATABASE_NAME = Contants.DATABASE_NAME;
     public static final String TABLE_USER_PROFILE = "user_profile_table";
 
@@ -31,8 +37,12 @@ public class DbHelper extends SQLiteOpenHelper {
 
     public DbHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
-
+        this.ctx = context;
+        this.DBNAME = DATABASE_NAME;
+        this.DBPATH = this.ctx.getDatabasePath(DBNAME).getAbsolutePath();
+        Log.e("Path 1", DBPATH);
     }
+
     @Override
     public void onCreate(SQLiteDatabase db) {
         String CREATE_USER_TABLE = " CREATE TABLE " + USER_DATA + " ( "
@@ -43,11 +53,11 @@ public class DbHelper extends SQLiteOpenHelper {
 
 
         String CREATE_USER_PROFILE_TABLE = " CREATE TABLE " + TABLE_USER_PROFILE + " ( "
-                + TIKTOK_ID + " TEXT PRIMARY KEY, "
-                + IMAGE_PROFILE + " TEXT, "
-                + IMAGE_VIDEO + " TEXT, "
+                + TIKTOK_EMAIL + " TEXT PRIMARY KEY, "
                 + TIKTOK_USER_NAME + " TEXT, "
-                + TIKTOK_USER_BIO + " TEXT " + ")";  //
+                + TIKTOK_CONTACT + " TEXT, "
+                + USER_PROFILE_PIC + " TEXT, "
+                + USER_BIO + " TEXT " + ")";  //
         db.execSQL(CREATE_USER_PROFILE_TABLE);
 
 
@@ -55,8 +65,10 @@ public class DbHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + USER_DATA);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USER_PROFILE);
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + USER_DATA);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_USER_PROFILE);
+        }
     }
 
 
@@ -98,7 +110,7 @@ public class DbHelper extends SQLiteOpenHelper {
         SQLiteDatabase db = this.getWritableDatabase();
 
         long i = db.insert(USER_DATA, null, values);
-        Log.d("insert",ob.getEmailAddress()+" "+ob.getUserName()+" "+ob.getUserPassword());
+        Log.d("insert", ob.getEmailAddress() + " " + ob.getUserName() + " " + ob.getUserPassword());
         db.close();
         return i > 0;
     }
@@ -111,11 +123,11 @@ public class DbHelper extends SQLiteOpenHelper {
     }
 
     //userData data
-    public LoginResponseData getUserData() {
+    public LoginResponseData getUserDataa() {
         String query = "Select * FROM userData";
         SQLiteDatabase db = this.getReadableDatabase();
         Cursor cursor = db.rawQuery(query, null);
-      //  LoginResponseData data = new LoginResponseData(EMAIL,NAME, PASSWORD);
+        //  LoginResponseData data = new LoginResponseData(EMAIL,NAME, PASSWORD);
         LoginResponseData data = new LoginResponseData();
         if (cursor.moveToFirst()) {
             cursor.moveToFirst();
@@ -132,14 +144,14 @@ public class DbHelper extends SQLiteOpenHelper {
     //userData data
     public LoginResponseData getUserDataByLoginId(String id) {
         String query = "Select * FROM " + USER_DATA + " WHERE " + EMAIL + " ='" + id + "'";
-       // String query = "Select * FROM userData WHERE id = " + id + " ";
+        // String query = "Select * FROM userData WHERE id = " + id + " ";
         SQLiteDatabase db = this.getReadableDatabase();
         Cursor cursor = db.rawQuery(query, null);
         LoginResponseData data = new LoginResponseData();
         if (cursor.moveToFirst()) {
             cursor.moveToFirst();
             populateUserformation(cursor, data);
-            Log.d("getLoginData",data.getEmailAddress()+" "+data.getUserPassword());
+            Log.d("getLoginData", data.getEmailAddress() + " " + data.getUserPassword());
             cursor.close();
         } else {
             data = null;
@@ -156,9 +168,9 @@ public class DbHelper extends SQLiteOpenHelper {
         values.put(PASSWORD, ob.getUserPassword());
         SQLiteDatabase db = this.getWritableDatabase();
         long i = 0;
-        i = db.update(USER_DATA, values, EMAIL+" = '" + ob.getEmailAddress() + "' ", null);
+        i = db.update(USER_DATA, values, EMAIL + " = '" + ob.getEmailAddress() + "' ", null);
 
-        Log.d("insert",ob.getEmailAddress()+" "+ob.getUserName()+" "+ob.getUserPassword());
+        Log.d("insert", ob.getEmailAddress() + " " + ob.getUserName() + " " + ob.getUserPassword());
         db.close();
         return i > 0;
     }
@@ -168,83 +180,143 @@ public class DbHelper extends SQLiteOpenHelper {
     public boolean deleteUserData(String id) {
         boolean result = false;
         SQLiteDatabase db = this.getWritableDatabase();
-        String query = EMAIL+" = " + id + " ";
+        String query = EMAIL + " = " + id + " ";
         db.delete(USER_DATA, query, null);
         db.close();
         return result;
     }
 
-
     //Get and set User profile data
 
     //***************************************UserProfileData*************************************//
 
-    public boolean upsertUserProfileData(UserProfileResponse ob) {
+    public boolean upsertUserData(UserProfileResponse ob) {
         boolean done = false;
         UserProfileResponse data = null;
-        if (ob.getUser_tiktok_id() != null) {
-            data = getUserProfileData(ob.getUser_tiktok_id());
+        if (ob.getEmail() != null) {
+            data = getUserDataById(ob.getEmail());
             if (data == null) {
-                done = insertUserProfileData(ob);
+                done = insertUserData(ob);
             } else {
-                done = updateUserProfileData(ob);
+                done = updateUserata(ob);
             }
         }
         return done;
     }
+    //id TEXT,helplineName TEXT,helplineNumber TEXT,helplineImage TEXT,helplineWebsite TEXT,description TEXT,)";
 
-    public boolean insertUserProfileData(UserProfileResponse ob) {
-        SQLiteDatabase db = this.getWritableDatabase();
+    private void populateUserData(Cursor cursor, UserProfileResponse ob) {
+        ob.setEmail(cursor.getString(0));
+        ob.setName(cursor.getString(1));
+        ob.setContact(cursor.getString(2));
+        ob.setProfile_Pic(cursor.getString(3));
+        ob.setDescription(cursor.getString(4));
+    }
+
+
+    public boolean insertUserData(UserProfileResponse ob) {
         ContentValues values = new ContentValues();
-        values.put(TIKTOK_ID, ob.getUser_tiktok_id());
-        values.put(IMAGE_PROFILE, ob.getImage_profile());
-        values.put(IMAGE_VIDEO, ob.getImage_video());
-        values.put(TIKTOK_USER_NAME, ob.getUser_Name());
-        values.put(TIKTOK_USER_BIO, ob.getUser_Bio());
+        values.put(TIKTOK_EMAIL, ob.getEmail());
+        values.put(TIKTOK_USER_NAME, ob.getName());
+        values.put(TIKTOK_CONTACT, ob.getContact());
+        values.put(USER_PROFILE_PIC, ob.getProfile_Pic());
+        values.put(USER_BIO, ob.getDescription());
+        SQLiteDatabase db = this.getWritableDatabase();
         long i = db.insert(TABLE_USER_PROFILE, null, values);
-        db.close();
-        Log.d("", "insertUserProfileData " + values);
-        return true;
-    }
-
-    public UserProfileResponse getUserProfileData(String tiktok_id) {
-        String query = "Select * FROM " + TABLE_USER_PROFILE + " WHERE " + TIKTOK_ID + " ='" + tiktok_id + "'";
-        // String deleteQuery = "DELETE FROM Students where StudentId='"+ id +"'";
-        SQLiteDatabase db = this.getReadableDatabase();
-        Cursor cursor = db.rawQuery(query, null);
-        UserProfileResponse ob = new UserProfileResponse();
-        if (cursor != null && cursor.moveToFirst()) {
-            ob.setUser_tiktok_id(cursor.getString(0));
-            ob.setImage_profile(cursor.getString(1));
-            ob.setImage_video(cursor.getString(2));
-            ob.setUser_Name(cursor.getString(3));
-            ob.setUser_Bio(cursor.getString(4));
-            cursor.moveToFirst();
-            cursor.close();
-        } else {
-            ob = null;
-        }
-        db.close();
-        return ob;
-    }
-
-    public boolean updateUserProfileData(UserProfileResponse ob) {
-        ContentValues values = new ContentValues();
-        values.put(TIKTOK_ID, ob.getUser_tiktok_id());
-        values.put(IMAGE_PROFILE, ob.getImage_profile());
-        values.put(IMAGE_VIDEO, ob.getImage_video());
-        values.put(TIKTOK_USER_NAME, ob.getUser_Name());
-        values.put(TIKTOK_USER_BIO, ob.getUser_Bio());
-        SQLiteDatabase db = this.getWritableDatabase();
-        long i = 0;
-        i = db.update(TABLE_USER_PROFILE, values, TIKTOK_ID + " = '" + ob.getUser_tiktok_id() + "' ", null);
+        Log.d("insertUserData", String.valueOf(values));
         db.close();
         return i > 0;
     }
 
-    public void deleteAllUserProfile() {
-        getWritableDatabase().delete(TABLE_USER_PROFILE, null, null);
 
+    public UserProfileResponse getUserData() {
+        String query = "Select * FROM " + TABLE_USER_PROFILE;
+        SQLiteDatabase db = this.getReadableDatabase();
+        Cursor cursor = db.rawQuery(query, null);
+        // UserList data = new UserList("address", "userMobile", "userName", "email", "profile", "latitude", "longitude", "distance","current_location");//
+        UserProfileResponse data = new UserProfileResponse();//
+        if (cursor.moveToFirst()) {
+            cursor.moveToFirst();
+            populateUserData(cursor, data);
+            cursor.close();
+        } else {
+            data = null;
+        }
+        db.close();
+        return data;
+    }
+
+    //get all Addresses list data........................
+    public List<UserProfileResponse> GetAllUserListData() {
+        List<UserProfileResponse> list = new ArrayList<UserProfileResponse>();
+        String query = "Select * FROM " + TABLE_USER_PROFILE;
+        SQLiteDatabase db = this.getReadableDatabase();
+        Cursor cursor = db.rawQuery(query, null);
+        if (cursor.moveToFirst()) {
+            while (cursor.isAfterLast() == false) {
+                //   UserList ob = new UserList("address", "userMobile", "userName", "email", "profile", "latitude", "longitude", "distance","current_location");//,"current_location"
+                UserProfileResponse ob = new UserProfileResponse();//,"current_location"
+                // populateUserData(cursor, ob);
+                ob.setEmail(cursor.getString(0));
+                ob.setName(cursor.getString(1));
+                ob.setContact(cursor.getString(2));
+                ob.setProfile_Pic(cursor.getString(3));
+                ob.setDescription(cursor.getString(4));
+                list.add(ob);
+                cursor.moveToNext();
+            }
+        }
+        db.close();
+        return list;
+    }
+
+    public UserProfileResponse getUserDataById(String email) {
+        String query = "Select * FROM " + TABLE_USER_PROFILE + " WHERE " + TIKTOK_EMAIL + " = '" + email + "'";
+        // String query = "Select * FROM " + TABLE_USER_PROFILE + " WHERE " + TIKTOK_EMAIL + " = " + email ;
+        SQLiteDatabase db = this.getReadableDatabase();
+        Cursor cursor = db.rawQuery(query, null);
+        //  UserList ob = new UserList("address", "userMobile", "userName", "email", "profile", "latitude", "longitude", "distance","current_location");//,"current_location"
+        UserProfileResponse ob = new UserProfileResponse();//,"current_location"
+        try {
+            if (cursor.moveToNext()) {
+                ob.setEmail(cursor.getString(0));
+                ob.setName(cursor.getString(1));
+                ob.setContact(cursor.getString(2));
+                ob.setProfile_Pic(cursor.getString(3));
+                ob.setDescription(cursor.getString(4));
+            } else {
+                ob = null;
+            }
+
+        } catch (Exception ex) {
+
+        } finally {
+            cursor.close();
+            db.close();
+        }
+        return ob;
+
+    }
+
+    //id TEXT,latitude TEXT,longitude TEXT,userName TEXT,userMobile TEXT,address TEXT)";
+    public boolean updateUserata(UserProfileResponse ob) {
+        ContentValues values = new ContentValues();
+        values.put(TIKTOK_EMAIL, ob.getEmail());
+        values.put(TIKTOK_USER_NAME, ob.getName());
+        values.put(TIKTOK_CONTACT, ob.getContact());
+        values.put(USER_PROFILE_PIC, ob.getProfile_Pic());
+        values.put(USER_BIO, ob.getDescription());
+        //  values.put("current_location", ob.getCurrent_location());
+        SQLiteDatabase db = this.getWritableDatabase();
+        long i = 0;
+        i = db.update(TABLE_USER_PROFILE, values, TIKTOK_EMAIL + " ='" + ob.getEmail() + "' ", null);
+        Log.d("updateUserata", String.valueOf(values));
+        db.close();
+        return i > 0;
+    }
+
+    public void deleteDeviceLatLongData() {
+        getWritableDatabase().delete(TABLE_USER_PROFILE, null, null);
     }
